requirements.txt
aioredis==2.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
async-timeout==5.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.123.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
packaging==25.0
protobuf==6.33.1
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
setuptools==80.9.0
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wheel==0.45.1
wrapt==1.17.3
zipp==3.23.0
